Add components ModelList and Pair
This commit adds two components that are useful when rendering models. A
`ModelList` and a `Pair`.

The `ModelList` is modelled after the `CollectionList` with some major
differences:

- It responds to model `change` events instead of `add` and `remove`.
- It expects a `componentFactory` function that takes a `key` and a
  `value` instead of an `item` and an `index`.

The default behavior is to call the `componentFactory` once for each
association in the model using alphabetical ordering of the keys. The
actual set of associations can be controlled using the options `include`
and `exclude`, both arrays of keys. The `include` option, if present,
overrides the alphabetical ordering. A key present in both `exclude` and
`include` will be excluded.

The `Pair` is just a grouping of two `Elem` instances and provided as a
convenient way to return a component from the component factory that
contains both a key and a value component.